+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - reader: AttrKeyIter and FileReader::iter_attr_keys for iterating through the keys of all the attributes
 - auxiliary: EnumField type for getting/setting enum values for auxiliary fields
 - record: Record::get_aux_enum_field to get enum label indices
+- writer: WriteOptions::aux_enum for adding enums to SLOW5 files
 
 
 ### Changed